Work in progress: convert SVG path to points.
For now, this is just placeholder proof-of-concept code.

I've been building out svglib.js with the eventual intention
of it becoming a well-organized library, but since Worker instances
don't have a window or a DOM parser, for now the easiest way to make
this work was to just pass the SVG path string back to the main window.

In order to handle asynchronous rendering, I needed to hack the
Worker message passing. Long-term, I would like to build my objects
by returning an object that implements a render method. The render
method would be responsible for emitting a "rendered" event once
the data was in place and ready to go, and would be attached to
the event. The base worker object would listen for events and
convert them to messages, to be passed back to the main window.

Or something along those lines, anyway. As I get deeper into the guts
of OpenJSCAD.js, I may come up with some other plan of attack.

All this code is garbage right now. But it's garbage that produces results,
so I don't want to lose it.